requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
feedparser==6.0.10
python-telegram-bot==20.6
slack-sdk==3.23.0
pyyaml==6.0.1
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
newspaper3k==0.2.8
pandas==2.1.3
aiohttp==3.9.0
asyncio==3.4.3
python-dateutil==2.8.2 